Adicionando funcionalidade de alteração do contato

+++ SPLMessage/src/br/com/message/facade/ContatoFacadeImpl.java
@@ -7,9 +7,12 @@ package br.com.message.facade;
 import java.util.Date;
 import java.util.List;
 
+import org.hibernate.cfg.beanvalidation.IntegrationException;
+
 import br.com.message.dao.ContatoDao;
 import br.com.message.dao.ContatoDaoImpl;
 import br.com.message.model.Contato;
+import br.com.message.model.Usuario;
 import br.com.message.util.DataStore;
 
 /**
@@ -21,12 +24,30 @@ public class ContatoFacadeImpl implements ContatoFacade {
 	private static final Integer ID_GRUPO_DEFAULT = 1;
 	
 	private ContatoDao contatoDao = new ContatoDaoImpl();
-
+	private UsuarioFacade usuarioFacade = new UsuarioFacadeImpl();
+	
 	//#if ${AdicionarContato} == "T"
 	@Override
 	public void inserir(Contato contato) {
-		contato.setIdGrupo(ID_GRUPO_DEFAULT);
+		Usuario c = usuarioFacade.findByEmail(contato.setEmailContato());
+		
+		//Valida se é um contato válido
+		if(c == null){
+			throw new IntegrationException("Não foi possível localizar o contato");
+		} else if(c.getEmail().equals(DataStore.getInstance().getUsuario().getEmail())){
+			throw new IntegrationException("Este contato não é válido");
+		}
+		
+		//Valida se o grupo foi selecionado
+		if(contato.getIdGrupo() == null){
+			contato.setIdGrupo(ID_GRUPO_DEFAULT);
+		}
+		
 		contato.setEmailUsuario(DataStore.getInstance().getUsuario().getEmail());
+		if(buscar(contato) != null){
+			throw new IntegrationException("Contato já cadastrado");
+		}
+		
 		contato.setDataInclusao(new Date());
 		this.contatoDao.inserir(contato);
 	}
@@ -34,8 +55,42 @@ public class ContatoFacadeImpl implements ContatoFacade {
 	
 	//#if ${RemoverContato} == "T"
 	@Override
-	public void remover(Contato contato) {
-		this.contatoDao.remover(contato);
+	public Usuario remover(Contato contato) {
+		Usuario c = usuarioFacade.findByEmail(contato.setEmailContato());
+		
+		//Valida se é um contato válido
+		if(c == null){
+			throw new IntegrationException("Não foi possível localizar o contato");
+		}
+		
+		if(buscar(contato) != null){
+			this.contatoDao.remover(contato);
+		} else {
+			throw new IntegrationException("Não foi possível localizar o contato");
+		}
+		
+		return c;
+	}
+	//#endif
+	
+	//#if ${AlterarContato} == "T"
+	@Override
+	public void update(Contato contato) {
+		Contato c = buscar(contato);
+		
+		//Valida se é um contato válido
+		if(c == null){
+			throw new IntegrationException("Não foi possível localizar o contato");
+		}
+		
+		//Valida se o grupo foi selecionado
+		if(contato.getIdGrupo() == null){
+			c.setIdGrupo(ID_GRUPO_DEFAULT);
+		} else {
+			c.setIdGrupo(contato.getIdGrupo());
+		}
+		
+		this.contatoDao.alterar(c);
 	}
 	//#endif
